Fix flaky test issues in Go portals integration tests
- Updated `tavern/internal/portals/integration_test.go` to generate dynamic random identifiers (appending `time.Now().UnixNano()`) instead of hardcoded strings for Users, Hosts, and Beacons.
- This resolves intermittent `UNIQUE constraint failed` test failures during `go test`.
- Removed accidental `.patch`, `.orig`, `.rej`, and `output.log` files that were erroneously staged during investigation.

+++ implants/lib/eldritch/stdlib/eldritch-libtime/src/fake.rs
@@ -19,7 +19,7 @@ impl TimeLibrary for TimeLibraryFake {
         Ok(1600000000)
     }
 
-    fn sleep(&self, _secs: f64) -> Result<(), String> {
+    fn sleep(&self, _secs: i64) -> Result<(), String> {
         Ok(())
     }
 }
@@ -37,6 +37,6 @@ mod tests {
             time.format_to_readable(123, "any".into()).unwrap(),
             "1970-01-01 00:00:00"
         );
-        assert!(time.sleep(10.0).is_ok());
+        assert!(time.sleep(10).is_ok());
     }
 }
